CLASSES-2961 When a group is marked for sync, set it as ready for sync right away
Previously we were relying on the GrouperSyncJob (running once a
minute) to pull out the members and mark the group for sync.  This
works fine when there *are* members, but it's possible for a group to
be created with an empty members list.  When this happened, the group
never synced.