CLASSES-3027 Prevent chrome from downloading recorded wav files
When previewing an announcement that contains an audio recording,
Chrome on OSX would download the wav file in addition to showing the
HTML5 player.  It turns out CKEditor was emitting two sets of tags: an
<audio> tag and an <object> tag.  The <object> tag was being ignored,
but was causing the download.

Since all modern browsers now support the <audio> tag, it seems that
we don't need the <object> one anyway.  I've removed it, and replaced
the "player 1 of 2" "player 2 of 2" placeholders with a generic one.