CLASSES-3380 Support rosters with inactive students
Previously, all users imported from rosters would be loaded with a
status of "active" in the course site in question.  This commit
changes that behavior to only activate users whose status is "active"
in the roster.  Otherwise, they're imported as inactive.

This new behavior can be disabled at runtime by setting the following
sakai.properties entry:

     nyu.rosters-enable-inactive-users=false