NOJIRA Fix query for calculating unread counts
This was joining to MFR_UNREAD_STATUS_T on message ID but not
accounting for messages that were moved to a different topic.  When
this happens, MFR_UNREAD_STATUS_T gets a second entry for the same
message, but with the new topic ID.

When counting, make sure the MFR_UNREAD_STATUS_T topicID *and*
messageID agree.